# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TypeRacer API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  email: string;
  username: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  email: string;
  /**
     * @minLength 3
     * @maxLength 20
     */
  username: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type GameDifficulty = typeof GameDifficulty[keyof typeof GameDifficulty];


export const GameDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Game {
  id: number;
  paragraph: string;
  difficulty: GameDifficulty;
  createdAt: string;
  updatedAt?: string;
}

export interface GameListResponse {
  data: Game[];
  total: number;
  page: number;
  limit: number;
}

export type GameInputDifficulty = typeof GameInputDifficulty[keyof typeof GameInputDifficulty];


export const GameInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface GameInput {
  /**
     * @minLength 80
     * @maxLength 500
     */
  paragraph: string;
  difficulty: GameInputDifficulty;
}

export type GamePatchDifficulty = typeof GamePatchDifficulty[keyof typeof GamePatchDifficulty];


export const GamePatchDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface GamePatch {
  /**
     * @minLength 80
     * @maxLength 500
     */
  paragraph?: string;
  difficulty?: GamePatchDifficulty;
}

export interface Result {
  id: number;
  userId: number;
  gameId: number;
  wpm: number;
  accuracy: number;
  timeTakenMs: number;
  createdAt: string;
}

export type ResultDetailDifficulty = typeof ResultDetailDifficulty[keyof typeof ResultDetailDifficulty];


export const ResultDetailDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface ResultDetail {
  id: number;
  userId: number;
  gameId: number;
  wpm: number;
  accuracy: number;
  timeTakenMs: number;
  createdAt: string;
  paragraph: string;
  difficulty: ResultDetailDifficulty;
  username: string;
  /** @nullable */
  personalBestWpm?: number | null;
}

export interface ResultHistoryResponse {
  data: ResultDetail[];
  total: number;
  page: number;
  limit: number;
}

export interface ResultInput {
  gameId: number;
  wpm: number;
  accuracy: number;
  timeTakenMs: number;
}

export type LeaderboardEntryDifficulty = typeof LeaderboardEntryDifficulty[keyof typeof LeaderboardEntryDifficulty];


export const LeaderboardEntryDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface LeaderboardEntry {
  rank: number;
  username: string;
  wpm: number;
  accuracy: number;
  difficulty: LeaderboardEntryDifficulty;
}

export interface UserStats {
  totalRaces: number;
  /** @nullable */
  bestWpm: number | null;
  /** @nullable */
  avgWpm: number | null;
  /** @nullable */
  avgAccuracy: number | null;
  recentResults: Result[];
}

export type GetRandomGameParams = {
difficulty?: GetRandomGameDifficulty;
};

export type GetRandomGameDifficulty = typeof GetRandomGameDifficulty[keyof typeof GetRandomGameDifficulty];


export const GetRandomGameDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export type ListGamesParams = {
page?: number;
limit?: number;
};

export type GetResultHistoryParams = {
page?: number;
limit?: number;
};

export type GetLeaderboardParams = {
difficulty?: GetLeaderboardDifficulty;
limit?: number;
};

export type GetLeaderboardDifficulty = typeof GetLeaderboardDifficulty[keyof typeof GetLeaderboardDifficulty];


export const GetLeaderboardDifficulty = {
  all: 'all',
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;