# lib/api-zod/src/generated/types/gameListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TypeRacer API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Game } from './game';

export interface GameListResponse {
  data: Game[];
  total: number;
  page: number;
  limit: number;
}