# lib/api-zod/src/generated/types/gamePatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TypeRacer API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GamePatchDifficulty } from './gamePatchDifficulty';

export interface GamePatch {
  /**
     * @minLength 80
     * @maxLength 500
     */
  paragraph?: string;
  difficulty?: GamePatchDifficulty;
}