# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TypeRacer API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Result } from './result';

export interface UserStats {
  totalRaces: number;
  /** @nullable */
  bestWpm: number | null;
  /** @nullable */
  avgWpm: number | null;
  /** @nullable */
  avgAccuracy: number | null;
  recentResults: Result[];
}